Create seperate routing folder
