+++ backend/src/main/kotlin/no/bekk/controllers/QuestionController.kt
@@ -2,21 +2,8 @@ package no.bekk.controllers
 
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
-import io.ktor.server.application.*
-import io.ktor.server.response.*
-import io.ktor.server.routing.*
 import no.bekk.model.internal.Question
 import java.io.File
-
-fun Route.questionRouting() {
-    val questionController = QuestionController()
-    route("/questions") {
-        get {
-            val questions = questionController.getTestQuestions()
-            call.respond(questions)
-        }
-    }
-}
 
 class QuestionController {
