+++ backend/src/main/kotlin/no/bekk/plugins/Routing.kt
@@ -10,10 +10,10 @@ import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.encodeToJsonElement
-import no.bekk.controllers.questionRouting
 import no.bekk.services.AirTableService
-import no.bekk.controllers.tableRouting
 import no.bekk.database.DatabaseRepository
+import no.bekk.routes.questionRouting
+import no.bekk.routes.tableRouting
 import java.sql.SQLException
 
 val airTableService = AirTableService()
